To ease the transition from Ant to Maven, read test files from classpath instead of from the default filesystem.
