+++ test/cobol/koopa/cobol/grammar/directives/test/fixed/FixedDirectivesTests.java
@@ -1,0 +1,12 @@
+package koopa.cobol.grammar.directives.test.fixed;
+
+import koopa.cobol.grammar.directives.test.DirectivesTests;
+import koopa.cobol.sources.SourceFormat;
+
+public class FixedDirectivesTests extends DirectivesTests {
+
+	@Override
+	protected SourceFormat getSourceFormat() {
+		return SourceFormat.FIXED;
+	}
+}
