+++ test/cobol/koopa/cobol/grammar/preprocessing/test/CobolPreprocessingGrammarTests.java
@@ -1,6 +1,5 @@
 package koopa.cobol.grammar.preprocessing.test;
 
-import java.io.File;
 import java.io.Reader;
 import java.io.StringReader;
 
@@ -11,7 +10,6 @@ import koopa.cobol.sources.SourceFormat;
 import koopa.core.grammars.Grammar;
 import koopa.core.sources.Source;
 import koopa.dsl.stage.runtime.GrammarTestSuite;
-import koopa.dsl.stage.util.StageUtil;
 
 public class CobolPreprocessingGrammarTests extends GrammarTestSuite {
 
@@ -19,12 +17,6 @@ public class CobolPreprocessingGrammarTests extends GrammarTestSuite {
 
 	public CobolPreprocessingGrammarTests() {
 		project.setDefaultFormat(SourceFormat.FREE);
-	}
-
-	@Override
-	public File[] getStageFiles() {
-		return new File("test/cobol/koopa/cobol/grammar/preprocessing/test/")
-				.listFiles(StageUtil.getFilenameFilter());
 	}
 
 	@Override
