+++ test/sql/koopa/sql/grammar/test/EmbeddedSQLGrammarTests.java
@@ -1,6 +1,5 @@
 package koopa.sql.grammar.test;
 
-import java.io.File;
 import java.io.Reader;
 import java.io.StringReader;
 
@@ -8,7 +7,6 @@ import koopa.core.grammars.Grammar;
 import koopa.core.sources.BasicTokens;
 import koopa.core.sources.Source;
 import koopa.dsl.stage.runtime.GrammarTestSuite;
-import koopa.dsl.stage.util.StageUtil;
 import koopa.sql.grammar.SQLGrammar;
 
 /**
@@ -18,12 +16,6 @@ import koopa.sql.grammar.SQLGrammar;
  * overall test suite defined here !</b>
  */
 public class EmbeddedSQLGrammarTests extends GrammarTestSuite {
-	@Override
-	public File[] getStageFiles() {
-		return new File("test/sql/koopa/sql/grammar/test/")
-				.listFiles(StageUtil.getFilenameFilter());
-	}
-
 	@Override
 	public Grammar getGrammar() {
 		return SQLGrammar.instance();
